# artifacts/api-server/src/lib/pdf-generator.ts
import { db } from "@workspace/db";
import { reportsTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { logger } from "./logger";
import path from "path";
import fs from "fs";

const workspaceRoot = process.cwd().endsWith(path.join("artifacts", "api-server"))
  ? path.resolve(process.cwd(), "../..")
  : process.cwd();

const reportsDir = path.resolve(workspaceRoot, "artifacts/api-server/reports");

interface Bug {
  id: number;
  title: string;
  description: string | null;
  severity: string;
  status: string;
}

interface AuditData {
  id: number;
  projectId: number;
  projectName: string | null;
  performanceScore: number | null;
  accessibilityScore: number | null;
  seoScore: number | null;
  bestPracticesScore: number | null;
  bugsFound: number;
  overallScore: number | null;
  startedAt: Date | null;
  completedAt: Date | null;
  durationMs: number | null;
  findings: Record<string, unknown> | null;
}

export async function generatePdfReport(
  reportId: number,
  audit: Record<string, unknown>,
  bugs: Bug[]
): Promise<void> {
  try {
    // Ensure reports directory exists
    if (!fs.existsSync(reportsDir)) {
      fs.mkdirSync(reportsDir, { recursive: true });
    }

    const fileName = `report-${reportId}-${Date.now()}.txt`;
    const filePath = path.join(reportsDir, fileName);

    const auditData = audit as unknown as AuditData;

    // Generate a text-based report (PDF generation would require pdfkit in production)
    const content = `
QA AUTOMATION PORTAL - AUDIT REPORT
====================================

Report ID: ${reportId}
Project: ${auditData.projectName || "Unknown"}
Audit Run ID: ${auditData.id}
Generated: ${new Date().toISOString()}

LIGHTHOUSE SCORES
-----------------
Performance:     ${auditData.performanceScore ?? "N/A"}/100
Accessibility:   ${auditData.accessibilityScore ?? "N/A"}/100
SEO:             ${auditData.seoScore ?? "N/A"}/100
Best Practices:  ${auditData.bestPracticesScore ?? "N/A"}/100
Overall Score:   ${auditData.overallScore ?? "N/A"}/100

SUMMARY
-------
Total Bugs Found: ${auditData.bugsFound}
Duration: ${auditData.durationMs ? `${Math.round(auditData.durationMs / 1000)}s` : "N/A"}
Completed: ${auditData.completedAt ? new Date(auditData.completedAt).toISOString() : "N/A"}

BUGS FOUND
----------
${bugs.length === 0 ? "No bugs found." : bugs.map((b, i) => `
${i + 1}. [${b.severity.toUpperCase()}] ${b.title}
   Status: ${b.status}
   ${b.description || ""}
`).join("\n")}

SEVERITY BREAKDOWN
------------------
Critical: ${bugs.filter(b => b.severity === "critical").length}
High:     ${bugs.filter(b => b.severity === "high").length}
Medium:   ${bugs.filter(b => b.severity === "medium").length}
Low:      ${bugs.filter(b => b.severity === "low").length}

RECOMMENDATIONS
---------------
${auditData.performanceScore && auditData.performanceScore < 70 ? "• Optimize performance: compress images, enable caching, minify JS/CSS\n" : ""}${auditData.accessibilityScore && auditData.accessibilityScore < 80 ? "• Improve accessibility: add ARIA labels, improve color contrast\n" : ""}${auditData.seoScore && auditData.seoScore < 75 ? "• Enhance SEO: improve meta tags, heading structure, page speed\n" : ""}${bugs.some(b => b.severity === "critical") ? "• Address critical bugs immediately before next release\n" : ""}
Report generated by QA Automation Portal
`;

    fs.writeFileSync(filePath, content, "utf-8");

    const fileUrl = `/api/reports/download/${fileName}`;

    await db.update(reportsTable).set({
      status: "ready",
      fileUrl,
    }).where(eq(reportsTable.id, reportId));

    logger.info({ reportId, filePath }, "Report generated");
  } catch (error) {
    logger.error({ reportId, error }, "Report generation failed");
    await db.update(reportsTable).set({ status: "failed" }).where(eq(reportsTable.id, reportId));
  }
}