# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Automation Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bugs:
    get:
      operationId: listBugs
      tags: [bugs]
      summary: List bugs
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: auditRunId
          in: query
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assignedToId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bugs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bug"
  /bugs/{id}:
    get:
      operationId: getBug
      tags: [bugs]
      summary: Get bug by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"
    patch:
      operationId: updateBug
      tags: [bugs]
      summary: Update bug status or details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BugUpdate"
      responses:
        "200":
          description: Updated bug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"
  /bugs/{id}/comments:
    get:
      operationId: listBugComments
      tags: [bugs]
      summary: List comments on a bug
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BugComment"
    post:
      operationId: createBugComment
      tags: [bugs]
      summary: Add a comment to a bug
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BugCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BugComment"

  # Screenshots
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/mark-all-read:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # Dashboard
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # Auth
...
    Bug:
      type: object
      required: [id, projectId, auditRunId, title, severity, status, priority, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        auditRunId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        severity:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [open, in_progress, resolved, ignored]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        resolutionNotes:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BugUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [open, in_progress, resolved, ignored]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assignedToId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        resolutionNotes:
          type: string
    BugComment:
      type: object
      required: [id, bugId, userId, content, createdAt]
      properties:
        id:
          type: integer
        bugId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    BugCommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # Screenshots
...
    Notification:
      type: object
      required: [id, userId, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [audit_completed, audit_failed, critical_issue, project_added, bug_assigned]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        relatedId:
          type: ["integer", "null"]
        relatedType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # Dashboard
...